Use intptr_t instead int to support x86_64 systems

+++ go-libucl.h
@@ -23,7 +23,7 @@ extern bool go_macro_call(int idx, ucl_object_t *arguments, char *data, int leng
 
 // Indirection that actually calls the Go macro handler.
 static inline bool _go_macro_handler(const unsigned char *data, size_t len, const ucl_object_t *arguments, void* ud) {
-    return go_macro_call((int)ud, (ucl_object_t *)arguments, (char*)data, (int)len);
+    return go_macro_call((intptr_t)ud, (ucl_object_t *)arguments, (char*)data, (int)len);
 }
 
 // Returns the ucl_macro_handler that we have, since we can't get this
@@ -35,7 +35,7 @@ static inline ucl_macro_handler _go_macro_handler_func() {
 // This just converts an int to a void*, because Go doesn't let us do that
 // and we use an int as the user data for registering macros.
 static inline void *_go_macro_index(int idx) {
-    return (void *)idx;
+    return (void *)(intptr_t)idx;
 }
 
 typedef struct ucl_schema_error ucl_schema_error_t;
